+++ src/Aks/Aks/Commands/NewAzureRmAks.cs
@@ -113,10 +113,6 @@ namespace Microsoft.Azure.Commands.Aks
         [Parameter(Mandatory = false, HelpMessage = "Docker bridge cidr used for building Kubernetes network.")]
         public string DockerBridgeCidr { get; set; }
 
-        [Parameter(Mandatory = false, HelpMessage = "Node pool labels used for building Kubernetes network.")]
-
-        public Hashtable NodePoolLabel { get; set; }
-
         [Parameter(Mandatory = false, HelpMessage = "Aks custom headers used for building Kubernetes network.")]
 
         public Hashtable AksCustomHeader { get; set; }